cleared day 07

+++ day07/day07.h
@@ -6,5 +6,5 @@ std::string INPUT_DAY_7 = "C:\\Users\\anblo\\Documents\\CodingProjects\\AoC_Uni\
 std::string INPUT_DAY_7_example = "C:\\Users\\anblo\\Documents\\CodingProjects\\AoC_Uni\\inputs\\day07_example.txt";
 
 namespace day07{
-    long long applyOps(const std::vector<int> &nums, int mask);
+    unsigned long long applyOps(const std::vector<unsigned long long> &nums, int mask);
 }
